Modifies the testing of FETools::Extrapolate.
This modification ensures that the compression failure of LinearAlgebra::distributed::Vector is properly detected by modifying the test triangulation with a simple inversion symmetry of the refinement pattern. It does not change the norms of solution vector evaluated during the testing since the tested function is invariant under the inversion symmetry. However, it introduces locally owned DoFs at interfaces between coarse locally owned cells and refined ghost cells, which is precisely the reason the compression failure arises.