Fixing the compression bug of FETools::extrapolate.
Due to the different compress(insert) behaviour of Petsc/Trilinos and
LinearAlgebra::distributed vectors, the previous version of the
extrapolate method was failing with the latter type of vector on
distributed simulations with locally owned DoF at the interface between
a locally owned coarse cell and a more refined neighbouring ghost cell.

This bug was not detected before by the fe_tools_extrapolate_*** tests
of the testsuite because the tested triangulation included no such DoFs.
A simple mirror inversion of the refinement pattern allows to detect the
bug without having to change the tested outputs (the extrapolated
function is invariant under the inversion symmmetry), and was also
included in this commit in addition to the fix.